+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - NEW: Added `alias_email` and `destination_email` to `EmailForward`
 - CHANGED: Deprecated `from` and `to` fields in `EmailForward`
+- CHANGE: Drop support for Java JDK 17
+- CHANGE: Add support for Java JDK 23
+- CHANGE: `DomainCollaborators` have been deprecated and will be removed in the next major version. Please use our Domain Access Control feature.
 
 ## 1.0.0